Fixed a compiler warning (and remove an unnessesary cast)
--HG--
extra : convert_revision : 37145f1b4ec49b432ea571f96562d44f4c5a908e

+++ pidgin/gtkimhtml.c
@@ -1974,7 +1974,7 @@ static void gtk_smiley_tree_remove (GtkSmileyTree     *tree,
 
 		pos = strchr (t->values->str, *x);
 		if (pos)
-			t = t->children [(unsigned int) pos - (unsigned int) t->values->str];
+			t = t->children [pos - t->values->str];
 		else
 			return;
